refactor(registration): temporary solution to fix tests
The effort involved in finding another solution would not be justifiable at this time, as the email confirmation solution is still being rebuilt. Therefore, a `return true` is the cheapest justifiable option in this case, especially since it will only be present in the snapshot temporarily for a day or two.

Signed-off-by: Marcus Fihlon <marcus@fihlon.swiss>