feat: Allow Local Users to Create New Community #274
For users of type LOCAL and with role USER, Add Community option will appear
in the Avatar menu upon the change. Clicking the button will take the user
to a form, where they can enter their desired Profile, Community Name and
Description.

The first two fields are required and can take up to 255 characters of data.
If the constraints are not met, the users will receive a validation on the fields,
as well as receive a notification at the bottom of the page, informing them of
the error.

If the data validation is satisfied, the user will be able to save the form. This
invovles creating the Community, as well as making the creator of the Community
the owner of the Community. A notification will be displayed at the bottom of the
page, prompting the user of the successful operation, as well as redirect user to
the Community list page.

Signed-off-by: Shamah M Zoha <shamah.zoha@gmail.com>