feat(admin): allow disabling community creation for local users #273
A new field has been added in the Edit Configuration page for the Admin
users, where they can decide whether the "Create Community" option would
appear in the Avatar menu for the Local users in the instance.

The field is labelled "Allow Community Creation", and the default value
defined as is "true", to allow for community creation out of the box.

Signed-off-by: Shamah M Zoha <shamah.zoha@gmail.com>