fix(ui): Fix Responsiveness for ConfirmationDialog
For ConfirmationDialog, custom CSS has been added, which ensures that once the
viewport goes under 600px, the dialog box will scale relative to the viewport
height and width. While the height should match 100% of the viewport height, the
width is set to match 90% of the viewport width, to allow for some padding.

This addresses issue #312.

Signed-off-by: Shamah M Zoha <shamah.zoha@gmail.com>